# Frontend/backend_openapi.json
{
    "openapi": "3.1.0",
    "info": {
      "title": "Valio Aimo Customer Portal API",
      "description": "Backend API for Zero-Fail Logistics Customer Portal",
      "version": "2.0.0"
    },
    "paths": {
...
      "/customer/claims/": {
        "get": {
          "tags": [
            "customer-claims"
          ],
          "summary": "Get Claims",
          "description": "Get all customer claims",
          "operationId": "get_claims_customer_claims__get",
          "responses": {
            "200": {
              "description": "Successful Response",
              "content": {
                "application/json": {
                  "schema": {
                    "items": {
                      "$ref": "#/components/schemas/ClaimWithDetails"
                    },
                    "type": "array",
                    "title": "Response Get Claims Customer Claims  Get"
                  }
                }
              }
            }
          },
          "security": [
            {
              "OAuth2PasswordBearer": []
            }
          ]
        },
        "post": {
          "tags": [
            "customer-claims"
          ],
          "summary": "Create Claim",
          "description": "Create a claim for a delivered order with multiple file uploads",
          "operationId": "create_claim_customer_claims__post",
          "responses": {
            "200": {
              "description": "Successful Response",
              "content": {
                "application/json": {
                  "schema": {
                    "$ref": "#/components/schemas/Claim"
                  }
                }
              }
            }
          },
          "security": [
            {
              "OAuth2PasswordBearer": []
            }
          ]
        }
      },
      "/customer/claims/{claim_id}": {
        "get": {
          "tags": [
            "customer-claims"
          ],
          "summary": "Get Claim",
          "description": "Get specific claim details",
          "operationId": "get_claim_customer_claims__claim_id__get",
          "security": [
            {
              "OAuth2PasswordBearer": []
            }
          ],
          "parameters": [
            {
              "name": "claim_id",
              "in": "path",
              "required": true,
              "schema": {
                "type": "string",
                "title": "Claim Id"
              }
            }
          ],
          "responses": {
            "200": {
              "description": "Successful Response",
              "content": {
                "application/json": {
                  "schema": {
                    "$ref": "#/components/schemas/ClaimWithDetails"
                  }
                }
              }
            },
            "422": {
              "description": "Validation Error",
              "content": {
                "application/json": {
                  "schema": {
                    "$ref": "#/components/schemas/HTTPValidationError"
                  }
                }
              }
            }
          }
        }
      },
      "/customer/claims/{claim_id}/messages": {
        "get": {
          "tags": [
            "customer-claims"
          ],
          "summary": "Get Claim Messages",
          "description": "Get messages for a claim",
          "operationId": "get_claim_messages_customer_claims__claim_id__messages_get",
          "security": [
            {
              "OAuth2PasswordBearer": []
            }
          ],
          "parameters": [
            {
              "name": "claim_id",
              "in": "path",
              "required": true,
              "schema": {
                "type": "string",
                "title": "Claim Id"
              }
            }
          ],
          "responses": {
            "200": {
              "description": "Successful Response",
              "content": {
                "application/json": {
                  "schema": {
                    "type": "array",
                    "items": {
                      "$ref": "#/components/schemas/Message"
                    },
                    "title": "Response Get Claim Messages Customer Claims  Claim Id  Messages Get"
                  }
                }
              }
            },
            "422": {
              "description": "Validation Error",
              "content": {
                "application/json": {
                  "schema": {
                    "$ref": "#/components/schemas/HTTPValidationError"
                  }
                }
              }
            }
          }
        },
        "post": {
          "tags": [
            "customer-claims"
          ],
          "summary": "Send Claim Message",
          "description": "Send message for a claim",
          "operationId": "send_claim_message_customer_claims__claim_id__messages_post",
          "security": [
            {
              "OAuth2PasswordBearer": []
            }
          ],
          "parameters": [
            {
              "name": "claim_id",
              "in": "path",
              "required": true,
              "schema": {
                "type": "string",
                "title": "Claim Id"
              }
            }
          ],
          "requestBody": {
            "required": true,
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/MessageCreate"
                }
              }
            }
          },
          "responses": {
            "200": {
              "description": "Successful Response",
              "content": {
                "application/json": {
                  "schema": {
                    "$ref": "#/components/schemas/Message"
                  }
                }
              }
            },
            "422": {
              "description": "Validation Error",
              "content": {
                "application/json": {
                  "schema": {
                    "$ref": "#/components/schemas/HTTPValidationError"
                  }
                }
              }
            }
          }
        }
      },
...
      "/admin/claims/": {
        "get": {
          "tags": [
            "admin-claims"
          ],
          "summary": "Get All Claims",
          "description": "Get all claims with optional filters",
          "operationId": "get_all_claims_admin_claims__get",
          "security": [
            {
              "OAuth2PasswordBearer": []
            }
          ],
          "parameters": [
            {
              "name": "skip",
              "in": "query",
              "required": false,
              "schema": {
                "type": "integer",
                "default": 0,
                "title": "Skip"
              }
            },
            {
              "name": "limit",
              "in": "query",
              "required": false,
              "schema": {
                "type": "integer",
                "default": 100,
                "title": "Limit"
              }
            },
            {
              "name": "status",
              "in": "query",
              "required": false,
              "schema": {
                "anyOf": [
                  {
                    "type": "string"
                  },
                  {
                    "type": "null"
                  }
                ],
                "title": "Status"
              }
            },
            {
              "name": "requires_manual_review",
              "in": "query",
              "required": false,
              "schema": {
                "anyOf": [
                  {
                    "type": "boolean"
                  },
                  {
                    "type": "null"
                  }
                ],
                "title": "Requires Manual Review"
              }
            }
          ],
          "responses": {
            "200": {
              "description": "Successful Response",
              "content": {
                "application/json": {
                  "schema": {
                    "type": "array",
                    "items": {
                      "$ref": "#/components/schemas/ClaimWithDetails"
                    },
                    "title": "Response Get All Claims Admin Claims  Get"
                  }
                }
              }
            },
            "422": {
              "description": "Validation Error",
              "content": {
                "application/json": {
                  "schema": {
                    "$ref": "#/components/schemas/HTTPValidationError"
                  }
                }
              }
            }
          }
        }
      },
...
      "/admin/claims/{claim_id}": {
        "get": {
          "tags": [
            "admin-claims"
          ],
          "summary": "Get Claim",
          "description": "Get specific claim details",
          "operationId": "get_claim_admin_claims__claim_id__get",
          "security": [
            {
              "OAuth2PasswordBearer": []
            }
          ],
          "parameters": [
            {
              "name": "claim_id",
              "in": "path",
              "required": true,
              "schema": {
                "type": "string",
                "title": "Claim Id"
              }
            }
          ],
          "responses": {
            "200": {
              "description": "Successful Response",
              "content": {
                "application/json": {
                  "schema": {
                    "$ref": "#/components/schemas/ClaimWithDetails"
                  }
                }
              }
            },
            "422": {
              "description": "Validation Error",
              "content": {
                "application/json": {
                  "schema": {
                    "$ref": "#/components/schemas/HTTPValidationError"
                  }
                }
              }
            }
          }
        }
      },
...
      "/admin/claims/{claim_id}/messages": {
        "get": {
          "tags": [
            "admin-claims"
          ],
          "summary": "Get Claim Messages",
          "description": "Get messages for a claim",
          "operationId": "get_claim_messages_admin_claims__claim_id__messages_get",
          "security": [
            {
              "OAuth2PasswordBearer": []
            }
          ],
          "parameters": [
            {
              "name": "claim_id",
              "in": "path",
              "required": true,
              "schema": {
                "type": "string",
                "title": "Claim Id"
              }
            }
          ],
          "responses": {
            "200": {
              "description": "Successful Response",
              "content": {
                "application/json": {
                  "schema": {
                    "type": "array",
                    "items": {
                      "$ref": "#/components/schemas/Message"
                    },
                    "title": "Response Get Claim Messages Admin Claims  Claim Id  Messages Get"
                  }
                }
              }
            },
            "422": {
              "description": "Validation Error",
              "content": {
                "application/json": {
                  "schema": {
                    "$ref": "#/components/schemas/HTTPValidationError"
                  }
                }
              }
            }
          }
        },
        "post": {
          "tags": [
            "admin-claims"
          ],
          "summary": "Send Claim Message",
          "description": "Send message for a claim",
          "operationId": "send_claim_message_admin_claims__claim_id__messages_post",
          "security": [
            {
              "OAuth2PasswordBearer": []
            }
          ],
          "parameters": [
            {
              "name": "claim_id",
              "in": "path",
              "required": true,
              "schema": {
                "type": "string",
                "title": "Claim Id"
              }
            }
          ],
          "requestBody": {
            "required": true,
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/MessageCreate"
                }
              }
            }
          },
          "responses": {
            "200": {
              "description": "Successful Response",
              "content": {
                "application/json": {
                  "schema": {
                    "$ref": "#/components/schemas/Message"
                  }
                }
              }
            },
            "422": {
              "description": "Validation Error",
              "content": {
                "application/json": {
                  "schema": {
                    "$ref": "#/components/schemas/HTTPValidationError"
                  }
                }
              }
            }
          }
        }
      },
      "/admin/customers/": {
        "post": {
          "tags": [
            "admin-customers"
          ],
          "summary": "Create Customer",
          "description": "Create a new customer",
          "operationId": "create_customer_admin_customers__post",
          "security": [
            {
              "OAuth2PasswordBearer": []
            }
          ],
          "requestBody": {
            "required": true,
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/CustomerCreate"
                }
              }
            }
          },
          "responses": {
            "200": {
              "description": "Successful Response",
              "content": {
                "application/json": {
                  "schema": {
                    "$ref": "#/components/schemas/Customer"
                  }
                }
              }
            },
            "422": {
              "description": "Validation Error",
              "content": {
                "application/json": {
                  "schema": {
                    "$ref": "#/components/schemas/HTTPValidationError"
                  }
                }
              }
            }
          }
        },
        "get": {
          "tags": [
            "admin-customers"
          ],
          "summary": "Get Customers",
          "description": "Get all customers",
          "operationId": "get_customers_admin_customers__get",
          "security": [
            {
              "OAuth2PasswordBearer": []
            }
          ],
          "parameters": [
            {
              "name": "skip",
              "in": "query",
              "required": false,
              "schema": {
                "type": "integer",
                "default": 0,
                "title": "Skip"
              }
            },
            {
              "name": "limit",
              "in": "query",
              "required": false,
              "schema": {
                "type": "integer",
                "default": 100,
                "title": "Limit"
              }
            }
          ],
          "responses": {
            "200": {
              "description": "Successful Response",
              "content": {
                "application/json": {
                  "schema": {
                    "type": "array",
                    "items": {
                      "$ref": "#/components/schemas/Customer"
                    },
                    "title": "Response Get Customers Admin Customers  Get"
                  }
                }
              }
            },
            "422": {
              "description": "Validation Error",
              "content": {
                "application/json": {
                  "schema": {
                    "$ref": "#/components/schemas/HTTPValidationError"
                  }
                }
              }
            }
          }
        }
      },
...
      "/admin/customers/{customer_id}": {
        "get": {
          "tags": [
            "admin-customers"
          ],
          "summary": "Get Customer",
          "description": "Get specific customer",
          "operationId": "get_customer_admin_customers__customer_id__get",
          "security": [
            {
              "OAuth2PasswordBearer": []
            }
          ],
          "parameters": [
            {
              "name": "customer_id",
              "in": "path",
              "required": true,
              "schema": {
                "type": "string",
                "title": "Customer Id"
              }
            }
          ],
          "responses": {
            "200": {
              "description": "Successful Response",
              "content": {
                "application/json": {
                  "schema": {
                    "$ref": "#/components/schemas/Customer"
                  }
                }
              }
            },
            "422": {
              "description": "Validation Error",
              "content": {
                "application/json": {
                  "schema": {
                    "$ref": "#/components/schemas/HTTPValidationError"
                  }
                }
              }
            }
          }
        }
      },
...
    },
    "components": {
      "schemas": {
...
        "HTTPValidationError": {
          "properties": {
            "detail": {
              "items": {
                "$ref": "#/components/schemas/ValidationError"
              },
              "type": "array",
              "title": "Detail"
            }
          },
          "type": "object",
          "title": "HTTPValidationError"
        },
...
        "Message": {
          "properties": {
            "message_id": {
              "type": "integer",
              "title": "Message Id"
            },
            "order_id": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Order Id"
            },
            "claim_id": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Claim Id"
            },
            "sender_type": {
              "$ref": "#/components/schemas/MessageSenderType"
            },
            "sender_id": {
              "type": "string",
              "title": "Sender Id"
            },
            "content": {
              "type": "string",
              "title": "Content"
            },
            "is_read": {
              "type": "boolean",
              "title": "Is Read"
            },
            "created_at": {
              "type": "string",
              "format": "date-time",
              "title": "Created At"
            }
          },
          "type": "object",
          "required": [
            "message_id",
            "sender_type",
            "sender_id",
            "content",
            "is_read",
            "created_at"
          ],
          "title": "Message"
        },
        "MessageCreate": {
          "properties": {
            "content": {
              "type": "string",
              "title": "Content"
            },
            "order_id": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Order Id"
            },
            "claim_id": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Claim Id"
            }
          },
          "type": "object",
          "required": [
            "content"
          ],
          "title": "MessageCreate"
        },
        "MessageSenderType": {
          "type": "string",
          "enum": [
            "customer",
            "admin",
            "ai"
          ],
          "title": "MessageSenderType"
        },
...
      },
      "securitySchemes": {
        "OAuth2PasswordBearer": {
          "type": "oauth2",
          "flows": {
            "password": {
              "scopes": {
  
              },
              "tokenUrl": "auth/login"
            }
          }
        }
      }
    }
  }